requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Medical imaging
nibabel>=3.2.0

# Computer vision & image processing
opencv-python>=4.5.0
imageio>=2.19.0
Pillow>=8.3.0
albumentations>=1.3.0

# Deep learning (PyTorch ecosystem)
torch>=1.11.0
torchvision>=0.12.0
tensorboard>=2.8.0

# Machine learning
scikit-learn>=1.0.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars
tqdm>=4.62.0

# Experiment tracking (optional)
wandb>=0.12.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Dimensionality reduction (optional for advanced EDA)
umap-learn>=0.5.0

# Development utilities
pathlib2>=2.3.0
